requirements.txt
fonttools==4.14.0
Pillow==7.2.0
python-telegram-bot==12.8